Move Lesson domain (viewset, serializers, permissions) to viewsets/lesson.py
