+++ kolibri/core/lessons/api_urls.py
@@ -2,7 +2,7 @@ from django.urls import include
 from django.urls import re_path
 from rest_framework import routers
 
-from .viewsets import LessonViewset
+from .viewsets.lesson import LessonViewset
 
 router = routers.SimpleRouter()
 router.register(r"lesson", LessonViewset, basename="lesson")
