+++ kolibri/core/lessons/viewsets/lesson.py
@@ -1,3 +1,6 @@
+from django_filters.rest_framework import DjangoFilterBackend
+from rest_framework.decorators import action
+from rest_framework.response import Response
 from rest_framework.serializers import BooleanField
 from rest_framework.serializers import CharField
 from rest_framework.serializers import ListField
@@ -6,16 +9,23 @@ from rest_framework.serializers import PrimaryKeyRelatedField
 from rest_framework.serializers import Serializer
 from rest_framework.serializers import ValidationError
 
-from .models import Lesson
-from .models import LessonAssignment
 from kolibri.core import error_constants
 from kolibri.core.api import HexUUIDField
 from kolibri.core.api import ValuesMethodField
+from kolibri.core.api import ValuesViewset
+from kolibri.core.auth.api import KolibriAuthPermissions
+from kolibri.core.auth.api import KolibriAuthPermissionsFilter
+from kolibri.core.auth.api import _ensure_raw_dict
 from kolibri.core.auth.constants.collection_kinds import ADHOCLEARNERSGROUP
 from kolibri.core.auth.models import Collection
 from kolibri.core.auth.models import FacilityUser
 from kolibri.core.auth.models import Membership
 from kolibri.core.auth.utils.users import create_adhoc_group_for_learners
+from kolibri.core.content.models import ContentNode
+from kolibri.core.content.utils.annotation import total_file_size
+from kolibri.core.lessons.models import Lesson
+from kolibri.core.lessons.models import LessonAssignment
+from kolibri.core.query import annotate_array_aggregate
 
 
 class ResourceSerializer(Serializer):
@@ -25,6 +35,8 @@ class ResourceSerializer(Serializer):
 
 
 class ClassroomSerializer(ModelSerializer):
+    # Use the FK attname column (a UUID) rather than "parent" to avoid
+    # traversing the relation — ValuesViewset fetches "collection__parent_id" directly.
     parent = CharField(source="parent_id", read_only=True)
 
     class Meta:
@@ -237,3 +249,74 @@ class LessonSerializer(ModelSerializer):
         return LessonAssignment.objects.create(
             assigned_by=self.context["request"].user, **params
         )
+
+
+class LessonPermissions(KolibriAuthPermissions):
+    # Overrides the default validator to sanitize the Lesson POST Payload
+    # before validation
+    def validator(self, request, view, datum):
+        model = view.get_serializer_class().Meta.model
+        validated_data = view.get_serializer().to_internal_value(
+            _ensure_raw_dict(datum)
+        )
+        # Cannot have create assignments without creating the Lesson first,
+        # so this doesn't try to validate the Lesson with a non-empty assignments list
+        validated_data.pop("assignments", [])
+        validated_data.pop("learner_ids", [])
+        return request.user.can_create(model, validated_data)
+
+
+class LessonViewset(ValuesViewset):
+    serializer_class = LessonSerializer
+    filter_backends = (KolibriAuthPermissionsFilter, DjangoFilterBackend)
+    filterset_fields = ("collection", "id")
+    permission_classes = (LessonPermissions,)
+    queryset = Lesson.objects.all().order_by("-date_created")
+
+    def consolidate(self, items, queryset):
+        if not items:
+            return items
+        adhoc_assignments = LessonAssignment.objects.filter(
+            lesson_id__in=[item["id"] for item in items],
+            collection__kind=ADHOCLEARNERSGROUP,
+        )
+        adhoc_assignments = annotate_array_aggregate(
+            adhoc_assignments,
+            # Filter excludes soft-deleted users from learner_ids (#444f7abf).
+            filter=FacilityUser.get_is_active_q("collection__membership"),
+            learner_ids="collection__membership__user_id",
+        )
+        adhoc_assignments = {
+            a["lesson"]: a
+            for a in adhoc_assignments.values("collection", "lesson", "learner_ids")
+        }
+        for item in items:
+            if item["id"] in adhoc_assignments:
+                adhoc_assignment = adhoc_assignments[item["id"]]
+                item["learner_ids"] = adhoc_assignment["learner_ids"]
+                item["assignments"] = [
+                    i
+                    for i in item["assignments"]
+                    if i != adhoc_assignment["collection"]
+                ]
+            else:
+                item["learner_ids"] = []
+            item["resources"] = item["resources"] or []
+        return items
+
+    def annotate_queryset(self, queryset):
+        return annotate_array_aggregate(
+            queryset, lesson_assignment_collections="lesson_assignments__collection"
+        )
+
+    @action(detail=False)
+    def size(self, request, **kwargs):
+        lessons = self.filter_queryset(self.get_queryset())
+        lessons_set = []
+        for lesson in lessons:
+            resource_nodes = ContentNode.objects.filter(
+                id__in=[r["contentnode_id"] for r in lesson.resources]
+            )
+            lessons_set.append({lesson.id: total_file_size(resource_nodes)})
+
+        return Response(lessons_set)
